+++ qupath-gui-fx/src/main/java/qupath/lib/gui/commands/ScriptInterpreter.java
@@ -89,6 +89,7 @@ import qupath.lib.gui.QuPathGUI;
 import qupath.fx.dialogs.Dialogs;
 import qupath.lib.gui.panes.ObjectTreeBrowser;
 import qupath.lib.gui.prefs.PathPrefs;
+import qupath.lib.gui.prefs.SystemMenubar;
 import qupath.lib.gui.scripting.QPEx;
 import qupath.lib.gui.tools.WebViews;
 import qupath.lib.images.ImageData;
@@ -536,8 +537,7 @@ class ScriptInterpreter {
 		paneMasterDetail.setDetailNode(tabPane);
 		pane.setTop(menuBar);
 		pane.setCenter(paneMasterDetail);
-		//			menuBar.setUseSystemMenuBar(true);
-		menuBar.useSystemMenuBarProperty().bindBidirectional(PathPrefs.useSystemMenubarProperty());
+		SystemMenubar.manageChildMenubar(menuBar);
 
 		stage.setScene(new Scene(pane, 800, 600));
 		textAreaInput.requestFocus();